move create_data_chunk_urls and implementation of get_channel_data to utils
