+++ tests/test_seerpy.py
@@ -245,74 +245,6 @@ class TestGetSegmentUrls:
         assert result.empty
 
 
-@mock.patch('seerpy.seerpy.SeerAuth', autospec=True)
-class TestCreateDataChunkUrls:
-
-    def test_success(self, seer_auth):
-        # setup
-        seer_auth.return_value.cookie = {'seer.sid': "cookie"}
-
-        # setup
-        meta_data = pd.read_csv(TEST_DATA_DIR / "study1_metadata_short_durations.csv", index_col=0)
-        segment_urls = pd.read_csv(TEST_DATA_DIR / "segment_urls_3.csv", index_col=0)
-
-        expected_result = pd.read_csv(TEST_DATA_DIR / "study1_data_chunk_urls.csv", index_col=0)
-
-        # run test
-        result = SeerConnect().create_data_chunk_urls(meta_data, segment_urls)
-
-        # check result
-        assert result.equals(expected_result)
-
-    def test_empty_input(self, seer_auth):
-        # setup
-        seer_auth.return_value.cookie = {'seer.sid': "cookie"}
-
-        # setup
-        meta_data = pd.read_csv(TEST_DATA_DIR / "empty_metadata.csv", index_col=0)
-        segment_urls = pd.read_csv(TEST_DATA_DIR / "empty_segment_urls.csv", index_col=0)
-
-        expected_result = pd.DataFrame(columns=['segments.id', 'dataChunks.url', 'dataChunks.time'])
-
-        # run test
-        result = SeerConnect().create_data_chunk_urls(meta_data, segment_urls)
-
-        # check result
-        assert result.equals(expected_result)
-
-    def test_empty_metadata(self, seer_auth):
-        # setup
-        seer_auth.return_value.cookie = {'seer.sid': "cookie"}
-
-        # setup
-        meta_data = pd.read_csv(TEST_DATA_DIR / "empty_metadata.csv", index_col=0)
-        segment_urls = pd.read_csv(TEST_DATA_DIR / "segment_urls_3.csv", index_col=0)
-
-        expected_result = pd.DataFrame(columns=['segments.id', 'dataChunks.url', 'dataChunks.time'])
-
-        # run test
-        result = SeerConnect().create_data_chunk_urls(meta_data, segment_urls)
-
-        # check result
-        assert result.equals(expected_result)
-
-    def test_empty_segments_urls(self, seer_auth):
-        # setup
-        seer_auth.return_value.cookie = {'seer.sid': "cookie"}
-
-        # setup
-        meta_data = pd.read_csv(TEST_DATA_DIR / "study1_metadata_short_durations.csv", index_col=0)
-        segment_urls = pd.read_csv(TEST_DATA_DIR / "empty_segment_urls.csv", index_col=0)
-
-        expected_result = pd.DataFrame(columns=['segments.id', 'dataChunks.url', 'dataChunks.time'])
-
-        # run test
-        result = SeerConnect().create_data_chunk_urls(meta_data, segment_urls)
-
-        # check result
-        assert result.equals(expected_result)
-
-
 @mock.patch('time.sleep', return_value=None)
 @mock.patch('seerpy.seerpy.GQLClient', autospec=True)
 @mock.patch('seerpy.seerpy.SeerAuth', autospec=True)
